Moved emitter files into their own package
